Move write_region to io
